Move non-Vectorizer funcs into separate matrix_utils module
