+++ textacy/vsm/__init__.py
@@ -1,1 +1,5 @@
+from .matrix_utils import (get_term_freqs, get_doc_freqs, get_inverse_doc_freqs,
+                           get_doc_lengths, get_information_content,
+                           apply_idf_weighting,
+                           filter_terms_by_df, filter_terms_by_ic)
 from .vectorizers import Vectorizer, GroupVectorizer
